# Small performance fix