Move static regexp to a global variable in order to avoid multiple
regexp pattern compilations.

Before fix:
```
BenchmarkPopulateQueryParameters-8   	    5000	    303218 ns/op	  217132 B/op	    2323 allocs/op
BenchmarkTemplateToSwaggerPath/with_JSON_names-8         	  200000	      9134 ns/op	    5848 B/op	     130 allocs/op
BenchmarkTemplateToSwaggerPath/without_JSON_names-8      	  200000	      9467 ns/op	    5486 B/op	     134 allocs/op
```

After fix:
```
BenchmarkPopulateQueryParameters-8   	   20000	     80937 ns/op	   10088 B/op	     345 allocs/op
BenchmarkTemplateToSwaggerPath/with_JSON_names-8         	  200000	      5846 ns/op	    2860 B/op	      99 allocs/op
BenchmarkTemplateToSwaggerPath/without_JSON_names-8      	  200000	      6039 ns/op	    2486 B/op	     103 allocs/op
```